+++ .github/workflows/test.yml
@@ -20,6 +20,7 @@ jobs:
       - name: Test no default features
         run: |
           nix develop -c cargo run --no-default-features --example instance
+          nix develop -c cargo run --no-default-features --example filler
           nix develop -c cargo run --no-default-features --example diff
           nix develop -c cargo run --no-default-features --example json
           nix develop -c cargo run --no-default-features --example rename-patch-struct
@@ -30,6 +31,7 @@ jobs:
       - name: Test with std features
         run: |
           nix develop -c cargo run --features=std --example instance
+          nix develop -c cargo run --features=std --example filler
           nix develop -c cargo run --features=std --example diff
           nix develop -c cargo run --features=std --example json
           nix develop -c cargo run --features=std --example rename-patch-struct
