+++ Cargo.toml
@@ -7,7 +7,7 @@ members = [
 
 [workspace.package]
 authors = ["Antonio Yang <yanganto@gmail.com>"]
-version = "0.8.7"
+version = "0.9.0"
 edition = "2021"
 categories = ["development-tools"]
 keywords = ["struct", "patch", "macro", "derive", "overlay"]
