+++ struct-patch/Cargo.toml
@@ -11,7 +11,7 @@ license.workspace = true
 readme.workspace = true
 
 [dependencies]
-struct-patch-derive = { version = "=0.8.7", path = "../struct-patch-derive" }
+struct-patch-derive = { version = "=0.9.0", path = "../struct-patch-derive" }
 
 [dev-dependencies]
 serde_json = "1.0"
